remap: Use new SHA1 server download link
This commit changes the download link for the Minecraft server to the
launcher link using SHA1. Because the new link contains the server's
SHA1 hash, this commit also replaces all MD5 checks for the server with
SHA1 checks.

Please note that this is not a security feature. Like MD5, SHA1 is not
considered safe nowadays. We merely use SHA1 because it's used in the
launcher.