world: Add doWaterFreezing game rule
This commit adds the `doWaterFreezing` game rule to prevent water from
freezing in cold biomes. As this behaviour is not affected by the
`randomTickSpeed` game rule, it was previously impossible to prevent
water from freezing without disabling `iceAndSnow` in PaperSpigot's
config entirely.

This commit needs the `GameRules.patch` in KigPaper's `nms-patches/`
directory as `GameRules.java` was not changed by (Paper)Spigot before
this patch.